requirements.txt
Flask==2.3.0
rdkit==2023.9.6
Pillow==10.0.0
openai==1.3.4
python-dotenv==0.21.0